[Madagascar] post feed (#262)
* update feed model

* add video info fossa

* update feed message

* update feed message

* update post feed

* update:

+++ fossa/src/main/java/com/keepreal/madagascar/fossa/model/PictureInfo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PictureInfo extends MediaInfo{
     private String url;
-    private Integer width;
-    private Integer height;
-    private Integer size;
+    private Long width;
+    private Long height;
+    private Long size;
 }
